Add initial README with project details and study background.
Reproducibility of the study "Assessment of alterations in histone modification function and guidance for death risk prediction in cervical cancer patients" by Zhao et al. (2022). This repository contains R code and workflows to reproduce the bioinformatics analysis conducted to evaluate the prognostic value of histone modifications in cervical cancer patients.